Separate ExpressionVisitor from FieldResolver
