+++ src/Nest/CommonAbstractions/Infer/Field/FieldResolver.cs
@@ -12,7 +12,7 @@ using System.Text;
 
 namespace Nest
 {
-	public class FieldResolver : ExpressionVisitor
+	public class FieldResolver
 	{
 		private readonly IConnectionSettingsValues _settings;
 
@@ -51,139 +51,24 @@ namespace Nest
 			string f;
 			if (this.Properties.TryGetValue(property, out f))
 				return f;
-			f = this.ResolveToLastToken(property.Expression, property.Property);
+			f = this.Resolve(property.Expression, property.Property, true);
 			this.Properties.TryAdd(property, f);
 			return f;
 		}
 
-		private string Resolve(Expression expression, MemberInfo member)
+		private string Resolve(Expression expression, MemberInfo member, bool toLastToken = false)
 		{
+			var visitor = new FieldExpressionVisitor(_settings);
 			var name = expression != null
-				? this.Resolve(expression)
+				? visitor.Resolve(expression, toLastToken)
 				: member != null
-					? this.Resolve(member)
+					? visitor.Resolve(member)
 					: null;
 
 			if (name == null)
 				throw new ArgumentException("Could not resolve a name from the given Expression or MemberInfo.");
 
 			return name;
-		}
-
-		private string ResolveToLastToken(Expression expression, MemberInfo member)
-		{
-			var name = expression != null
-				? this.ResolveToLastToken(expression)
-				: member != null
-					? this.Resolve(member)
-					: null;
-
-			if (name == null)
-				throw new ArgumentException("Could not resolve a name from the given Expression or MemberInfo.");
-
-			return name;
-		}
-
-		private string Resolve(MemberInfo info)
-		{
-			if (info == null)
-				return null;
-
-			var name = info.Name;
-
-			IPropertyMapping propertyMapping = null;
-			if (this._settings.PropertyMappings.TryGetValue(info, out propertyMapping))
-				return propertyMapping.Name;
-
-			var att = ElasticsearchPropertyAttribute.From(info);
-			if (att != null && !att.Name.IsNullOrEmpty())
-				return att.Name;
-
-			return _settings.Serializer?.CreatePropertyName(info) ?? _settings.DefaultFieldNameInferrer(name);
-		}
-
-		private string Resolve(Expression expression)
-		{
-			Stack = new Stack<string>();
-			Visit(expression);
-			return Stack
-				.Aggregate(
-					new StringBuilder(),
-					(sb, name) =>
-					(sb.Length > 0 ? sb.Append(".") : sb).Append(name))
-				.ToString();
-		}
-
-		private string ResolveToLastToken(Expression expression)
-		{
-			Stack = new Stack<string>();
-			Visit(expression);
-			return Stack.Last();
-		}
-
-		protected override Expression VisitMember(MemberExpression expression)
-		{
-			if (Stack == null) return base.VisitMember(expression);
-			var resolvedName = this.Resolve(expression.Member);
-			Stack.Push(resolvedName);
-			return base.VisitMember(expression);
-		}
-
-		protected override Expression VisitMethodCall(MethodCallExpression methodCall)
-		{
-			if (methodCall.Method.Name == "Suffix" && methodCall.Arguments.Any())
-			{
-				VisitConstantOrVariable(methodCall, Stack);
-				var callingMember = new ReadOnlyCollection<Expression>(
-					new List<Expression> { { methodCall.Arguments.First() } }
-				);
-				base.Visit(callingMember);
-				return methodCall;
-			}
-			else if (methodCall.Method.Name == "get_Item" && methodCall.Arguments.Any())
-			{
-				var t = methodCall.Object.Type;
-				var isDict =
-					typeof(IDictionary).IsAssignableFrom(t)
-					|| typeof(IDictionary<,>).IsAssignableFrom(t)
-					|| (t.IsGeneric() && t.GetGenericTypeDefinition() == typeof(IDictionary<,>));
-
-				if (!isDict)
-				{
-					return base.VisitMethodCall(methodCall);
-				}
-				VisitConstantOrVariable(methodCall, Stack);
-				Visit(methodCall.Object);
-				return methodCall;
-			}
-			else if (IsLinqOperator(methodCall.Method))
-			{
-				for (int i = 1; i < methodCall.Arguments.Count; i++)
-				{
-					Visit(methodCall.Arguments[i]);
-				}
-				Visit(methodCall.Arguments[0]);
-				return methodCall;
-			}
-			return base.VisitMethodCall(methodCall);
-		}
-
-		private static void VisitConstantOrVariable(MethodCallExpression methodCall, Stack<string> stack)
-		{
-			var lastArg = methodCall.Arguments.Last();
-			var constantExpression = lastArg as ConstantExpression;
-			var value = constantExpression != null
-				? constantExpression.Value.ToString()
-				: Expression.Lambda(lastArg).Compile().DynamicInvoke().ToString();
-			stack.Push(value);
-		}
-
-		private static bool IsLinqOperator(MethodInfo methodInfo)
-		{
-			if (methodInfo.DeclaringType != typeof(Queryable) && methodInfo.DeclaringType != typeof(Enumerable))
-				return false;
-
-			return methodInfo.GetCustomAttribute<ExtensionAttribute>() != null;
 		}
 	}
 }
